Add instructions for accessing the Rails server within Vagrant
Rails 4.2... doesn't recognize servers run in vagrant as being run locally.
This causes the regular `rails server` command to fail to load the application
on localhost:3000. We could create a patch to fix this and ask student to add it
to their new apps but that seems a little too tricky for new programmers. We
could also create a script that students could run each time before starting the
server, but that would most likely be very confusing. Having students simply
bind to localhost has no future ramifications and is the easiest solution.